requirements.txt
openai>=1.40.0
google-genai>=0.4.0
rich>=13.0.0
prompt_toolkit>=3.0.0
flask>=3.0.0

tavily-python>=0.5.0
requests>=2.31.0
